+++ engines.py
@@ -1,0 +1,27 @@
+__ENGINES_JSON__ = {
+    "LOCAL": {
+        "engine": "postgresql://tdobbins:tdobbins@localhost:5432/phTeam-allpayer",
+        "caution_level": "warning",
+        "order": 0
+    },
+    "DEV": {
+        "engine": "",
+        "caution_level": "warning",
+        "order": 1
+    },
+    "PROD": {
+        "engine": "",
+        "caution_level": "danger",
+        "order": 2
+    },
+    "ALLPAYER": {
+        "engine": "",
+        "caution_level": "danger",
+        "order": 3
+    },
+    "DATALAKE":{
+    	"engine": "",
+    	"caution_level": "danger",
+    	"order": 4
+    }
+}
